+++ packages/meeting_place_chat/lib/src/chat/individual/individual_didcomm_chat_sdk.dart
@@ -245,7 +245,7 @@ class IndividualDidcommChatSDK extends BaseChatSDK
   }
 
   @override
-  Future<Uint8List> downloadMedia(ChatAttachment attachment) {
+  Future<Uint8List> downloadMedia(Message message) {
     throw UnsupportedError(
       'Hosted media download is not supported on DIDComm chat sessions',
     );